Auto change lead status.
- Move mutation test to seperate file for lead.
- Change Lead Status Enum values.
