+++ apps/entry/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class EntryConfig(AppConfig):
     name = 'entry'
+
+    def ready(self):
+        import entry.receivers  # noqa
